Further refinements for LAMMPS
I discovered that simple if statements can be used to force values in a
combox and can also disable it.

The LAMMPS documentation seems understandable enough that I could
possibly support most of the functionality.  I'm not sure if I want to
think about a "basic" and "advanced" set of panes.  Using the website
below I can at least build the required constraints on the input
keywords.

http://lammps.sandia.gov/doc/Section_commands.html