The start of fhi-aims input generator
This input generator is initially a copy of the quantum espresso dialog.

Important differences between this generator and many others is that it
will write only the control.in and geometry.in files.  The user is not
given the option to change the names.  This should be a convience for
the user.  I still need to figure out how prompt the user about
replacing said files if already present.

The geometry.in file should now be correct.  Keywords in control.in
should be very simple.

Change-Id: Idccf03dbb530ef99faf4ba316a4b85df74831c93